requirements.txt
streamlit==1.36.0
google-cloud-aiplatform==1.60.0
redis==5.0.1
requests==2.32.3
beautifulsoup4==4.12.3
Unidecode==1.3.8
python-dotenv==1.0.1
